compiler unit tests
