+++ unit_tests/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "unit_tests",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "jest"
+  },
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "jest": "^29.7.0"
+  },
+  "dependencies": {
+    "jsonfile": "^6.1.0"
+  }
+}
